refactor(db): split reality_portal repository into cohesive sub-modules
Move-only split of the 2176-line repositories/reality_portal.rs monolith
into a reality_portal/ directory (same pattern as document/ #1683,
subscription/ #1796, sensor/ #1810, rental/ #2100):

- mod.rs      — struct + new(), module docs (pool field pub(super))
- searches.rs — saved-search alert engine, favorites, favorite-alert worker
- agencies.rs — agency CRUD, branding, members & invitations
- realtors.rs — realtor profiles
- inquiries.rs — listing inquiries & responses
- listings.rs — portal listing CRUD, views & analytics
- imports.rs  — import jobs & feed subscriptions
- tests.rs    — cross-realtor IDOR-guard integration tests

Method bodies, SQL, and signatures are unchanged (verified by multiset
line diff: only doc/mod-plumbing/import/visibility lines differ). Public
API preserved via the existing repositories/mod.rs re-export. Baseline
entry in rls-pool-field-baseline.txt moved to the reality_portal/mod.rs
directory-module key.

Re-mint of PR #1812 (closed by stale-review SLA after an infra-flake CI
failure — rustup install exit 126), rebased onto current dev including
the search-alert drainer and favorite-alert worker additions.

Closes #2152
